// src/auth/types.ts
export type ActionMapType<M extends { [index: string]: any }> = {
  [Key in keyof M]: M[Key] extends undefined
    ? {
        type: Key;
      }
    : {
        type: Key;
        payload: M[Key];
      };
};

export type AuthUserType = null | Record<string, any>;

export type AuthStateType = {
  status?: string;
  loading: boolean;
  user: AuthUserType;
};
// ...
export enum ESocialProvider {
  azureAD = 'AzureAD',
  google = 'Google',
  amazon = 'LoginWithAmazon',
  facebook = 'Facebook',
}